transaction api added

+++ src/controllers/categoryController.ts
@@ -1,9 +1,10 @@
 
-import { eq } from "drizzle-orm";
+import { eq ,and } from "drizzle-orm";
 import { db } from "../db";
 import { categories } from "../db/schema/category";
 import { AuthRequest } from "../middleware/auth";
 import { Request, Response } from "express";
+import { CategorySyncBody } from "../validators/catValidator";
 
 
 
@@ -15,7 +16,7 @@ export const getCategoryList = async (req:AuthRequest,res:Response)=>{
         if(!req.user) return res.status(401).json({success:false,msg:"Unauthorized"});
 
         const userCategories = await db.select().from(categories).where(eq(categories.userId,req.user.id));
-        return res.json({ success: true, categories: userCategories });
+        return res.json({ success: true, msg:"user categories retrived successfuly" ,categories: userCategories });
 
 
     } catch (err) {
@@ -29,14 +30,17 @@ export const getCategoryList = async (req:AuthRequest,res:Response)=>{
 export const syncCategories = async (req: AuthRequest, res: Response) => {
   try {
  
+      if(!req.user) return res.status(401).json({success:false,msg:"Unauthorized"});
 
-    const { categories: incomingCategories } = req.body;
-    const userId = req.user!.id;
+    const { categories: incomingCategories } = req.body as CategorySyncBody;
+    const userId = req.user?.id;
+
+
 
     for (const cat of incomingCategories) {
       if (cat.isDeleted) {
        
-        await db.delete(categories).where(eq(categories.id, cat.catId));
+        await db.delete(categories).where(and(eq(categories.id, cat.catId),eq(categories.userId,userId)));
         continue;
       }
 
@@ -49,7 +53,7 @@ export const syncCategories = async (req: AuthRequest, res: Response) => {
         
            updatedAt: new Date(),
         })
-        .where(eq(categories.id, cat.catId))
+        .where(and(eq(categories.id, cat.catId),eq(categories.userId,userId)))
         .returning();
 
       if (updated.length === 0) {
